requirements.txt
streamlit
langchain==0.1.10
langchain-openai==0.0.8
langchain-community==0.0.25
langchain-text-splitters==0.0.1
langchain-core==0.1.53
pypdf2
python-dotenv
faiss-cpu
openai
tiktoken
sentencepiece
sentence_transformers
cohere
langchain-cohere